pyproject.toml
[tool.poetry]
name = "python-drf"
version = "0.1.0"
description = ""
authors = ["Andrii-77 <andrijsmaga@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.7"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"



